// src/app/shared/controls/customs/buttons/shared/types/button-types.ts
export type TypeButton = "button" | "submit";
export type MatButtonType =
  | "mat-button"
  | "mat-raised-button"
  | "mat-stroked-button"
  | "mat-flat-button"
  | "mat-fab"
  | "mat-mini-fab"
  | "none";

export type TypeButtonClass = 
  | "btn-normal-red"
  | "btn-normal"
  | "btn-submit"
  | "btn-close"
  | "btn-disabled"
  | "btn-facebook"
  | "btn-google"
  | "none"